XWIKI-15988: Upgrade tests sometimes don't start fast enough

+++ xwiki-platform-core/xwiki-platform-test/xwiki-platform-test-integration/src/main/java/org/xwiki/test/integration/XWikiWatchdog.java
@@ -45,7 +45,7 @@ public class XWikiWatchdog
      *         allowing the caller to know if the XWiki instance is up or not
      * @throws Exception if a thread has interrupted the thread in which this code is executing
      */
-    public WatchdogResponse isXWikiStarted(String url, int timeout) throws Exception
+    public WatchdogResponse isXWikiStarted(String url, long timeout) throws Exception
     {
         HttpClient client = new HttpClient();
